Resolving circular dependency issue and rearranging a few things
