+++ app/api/deps.py
@@ -1,5 +1,4 @@
-from datetime import datetime, timedelta
-from typing import AsyncIterator, Optional
+from typing import AsyncIterator
 
 import app.crud.user as crud
 import app.schemas.user as schemas
@@ -9,25 +8,10 @@ from app.schemas.token import TokenData
 from fastapi import Depends, HTTPException
 from fastapi.security import OAuth2PasswordBearer
 from jose import JWTError, jwt
-from passlib.context import CryptContext
-from pydantic import BaseModel
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette import status
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl=settings.TOKEN_URL)
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-class AuthRequest(BaseModel):
-    key: str
-
-
-def verify_password(plain, hashed):
-    return pwd_context.verify(plain, hashed)
-
-
-def get_password_hash(plain):
-    return pwd_context.hash(plain)
 
 
 async def get_session() -> AsyncIterator[AsyncSession]:
@@ -38,20 +22,6 @@ async def get_session() -> AsyncIterator[AsyncSession]:
             raise e
         finally:
             await session.close()
-
-
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    to_encode = data.copy()
-
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(days=7)  # 7 day token expiration by default
-
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
-
-    return encoded_jwt
 
 
 async def get_current_user(session: AsyncSession = Depends(get_session), token: str = Depends(oauth2_scheme)):
