+++ app/crud/user.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from app.crud.base import CrudBase
-from app.db.utils import get_password_hash, verify_password
+from app.core.security import get_password_hash, verify_password
 from app.models.user import User
 from app.schemas.user import UserCreate, UserUpdate
 from sqlalchemy import select
