+++ docker-compose.yaml
@@ -14,13 +14,6 @@ services:
         command: uvicorn app.main:app --reload --host 0.0.0.0 --port 8000
         depends_on:
             - database
-        deploy:
-            resources:
-                reservations:
-                    devices:
-                        - driver: nvidia
-                          count: 1
-                          capabilities: [ gpu ]
 
     database:
         image: postgres:14-bullseye
